requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.5
django-crontab==0.7.1
django-templated-mail==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
flake8==3.8.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
prompt-toolkit==3.0.14
psycopg2==2.8.6
psycopg2-binary>=2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
